p_wave_detector_load_balance/requirements.txt
kafka-python
obspy
numpy
tensorflow==2.10.0
# tensorflow
fastapi
uvicorn